chore(clerk-js): Use shared layout for alternative methods
This change enables us to merge the 2 components AlternativeMethods and ForgotPasswordAlternativeMethods into a single one by adding one external prop which dictates which part of the ui to be visible to indicate the Reset password flow
