+++ packages/clerk-js/src/ui/elements/contexts/FlowMetadataContext.tsx
@@ -23,6 +23,7 @@ type FlowMetadata = {
     | 'emailLinkVerify'
     | 'emailLinkStatus'
     | 'alternativeMethods'
+    | 'forgotPasswordMethods'
     | 'havingTrouble'
     | 'ssoCallback'
     | 'popover'
